Added more README.md files and refactored common functions into src/app_utils
